Add ref documentation for Postgres connection string format
This change adds a pointer to Postgres documentation that defines
the format of a Postgres connection string. This is for the case
where the chart value `database.url` is being set explicitly
to a Postgres connection string for an external (i.e. not provided
by the Helm chart) Postgres database.

Addresses Issue #116