Fixes Jenkins master branch 'InvalidImageName' failures
This change fixes master branch build failures. Tests are failing
for both KinD GitHub Actions and the tests on GCP.

The root cause is that the Helm template for the Helm test container
(`conjur-oss/templates/test/test-simple-install.yaml`) was not
updated with the new values schema for the Conjur image.
For example, the template should use `.Values.image.kubernetes.repository` or
`.Values.image.openshift.repository` instead of `.Values.image.repository`.

This change also updates the Configuration table in the Chart's
README.md to reflect the new schema for the Conjur image.

Addresses Issue #122