Removes OCP specific image repository and tag values
This change eliminates the OpenShift specific image repository and tag
values from the chart. This is being done to revert breaking changes
in chart values that added separate image repository/tag values
for Kubernetes vs OpenShift platforms. This change combines the
various Kubernetes / Openshift values into common values:

- Conjur repository
- Conjur tag
- Postgres repository
- Postgres tag
- Nginx repository
- Nginx tag

The defaults for these values will be the Kubernetes platform default
values. For deploying on OpenShift, the user will need to explicitly
set these to values that are appropriate for OpenShift.

The default values to use for OpenShift are documented in the chart's
README.md and in the `values.yaml` file.