Fix undefined docker env var errors for KinD example
Due to recent changes to the conjurdemos/kubernetes-conjur-demo repo made
the Docker environment variables DOCKER_REGISTRY_URL and DOCKER_REGISTRY_PATH
are now required when running the scripts in the
conjurdemos/kubernetes-conjur-demo. These are used to derive appropriate
paths when the scripts push/pull Docker images to/from a registry.

This fix adds an appropriate value for DOCKER_REGISTRY_PATH for the
kubernetes-in-docker application deployment example, which uses a
private local registry for Docker pushes/pulls.