The there is an equivalent environmental variable to CONJUR_NAMESPACE (#161)
in other projects called CONJUR_NAMESPACE_NAME. The purpose is to set
the Conjur OSS namespace name in both cases. We need to check if either
one exists when setting CONJUR_NAMESPACE. If both CONJUR_NAMESPACE and
CONJUR_NAMESPACE_NAME are set we assume that CONJUR_NAMESPACE_NAME is the
preferred name and will use that.