Eliminates CONJUR_NAMESPACE undefined errors in KinD example
Currently, if you run the Kubernetes-in-Docker (KinD) example in
'examples/kubernetes-in-docker', the scripts generate an error for
CONJUR_NAMESPACE being an undefined environment variable. This is fixed by
reversing the order in which CONJUR_NAMESPACE and CONJUR_NAMESPACE_NAME
environment variables are evaluated.

This also eliminates a minor, benign error message about a non-existent
"conjur-oss" Namespace when the scripts check to see if this Namespace
exists and it does not exist.